+++ vehicle.h
@@ -5,8 +5,8 @@
 class vehicle : public obstacle
 {
 	private:
-		int* rear_centroid_x;
-		int* rear_centroid_y;
+		double* rear_centroid_x;
+		double* rear_centroid_y;
 		int vehicle_center_x, vehicle_center_y;
 		double orientation;
 
@@ -15,13 +15,13 @@ class vehicle : public obstacle
 		void update_position_orientation();
 
 		// Update component centroid locations, recalculate vehicle center and orientation
-		void update_position_orientation(int &front_centroid_x, int &front_centroid_y, int &back_centroid_x, int &back_centroid_y);
+		void update_position_orientation(double &front_centroid_x, double &front_centroid_y, double &back_centroid_x, double &back_centroid_y);
 
 		// Default Constructor places vehicle outside of world map
 		vehicle();
 
 		// Preferred Constructor places vehicle where it is found on the world map
-		vehicle(int *front_centroid_x, int *front_centroid_y, int *back_centroid_x, int *back_centroid_y, int vehicle_diameter, int vehicle_keep_out);
+		vehicle(double *front_centroid_x, double *front_centroid_y, double *back_centroid_x, double *back_centroid_y, int vehicle_diameter, int vehicle_keep_out);
 
 		// Getters
 
@@ -33,7 +33,7 @@ class vehicle : public obstacle
 
 		// Setters
 
-		void set_rear_center_x(int* rear_centroid_x);
-		void set_rear_center_y(int* rear_centroid_y);
+		void set_rear_center_x(double* rear_centroid_x);
+		void set_rear_center_y(double* rear_centroid_y);
 };
